+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2021-04-20
+- increased pdf creator library dependency to 0.3
+- added support for Dompdf (translations and custom debug implementation)
+- added support for feature support check to hide functionality not provided by a pdf creator class
+
 ## [0.2.1] - 2021-04-06
 - fixed missing syndication type translation